I fixed the error by changing int to LocalDateTime, as well as replacing keys on DB with Long and converting int to its wrapper class, specifically Integer

+++ src/main/java/com/example/home_work_49/models/ContactInfo.java
@@ -1,8 +1,8 @@
 package com.example.home_work_49.models;
 
 public class ContactInfo {
-    private int id;
-    private int typeId;
-    private int resumeId;
+    private Long id;
+    private Long typeId;
+    private Long resumeId;
     private String value;
 }
