+++ src/main/java/com/example/home_work_49/models/User.java
@@ -3,10 +3,10 @@ package com.example.home_work_49.models;
 import lombok.Getter;
 
 public class User {
-    private int id;
+    private Long id;
     private String name;
     private String surname;
-    private int age;
+    private Integer age;
     private String email;
     private String password;
     private String phoneNumber;
